// autogpt_platform/frontend/src/app/api/__generated__/models/needLoginResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { NeedLoginResponseSessionId } from "./needLoginResponseSessionId";
import type { NeedLoginResponseAgentInfo } from "./needLoginResponseAgentInfo";

/**
 * Response when login is needed.
 */
export interface NeedLoginResponse {
  type?: ResponseType;
  message: string;
  session_id?: NeedLoginResponseSessionId;
  agent_info?: NeedLoginResponseAgentInfo;
}