// autogpt_platform/frontend/src/app/api/__generated__/models/setupRequirementsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { SetupRequirementsResponseSessionId } from "./setupRequirementsResponseSessionId";
import type { SetupInfo } from "./setupInfo";
import type { SetupRequirementsResponseGraphId } from "./setupRequirementsResponseGraphId";
import type { SetupRequirementsResponseGraphVersion } from "./setupRequirementsResponseGraphVersion";

/**
 * Response for validate action.
 */
export interface SetupRequirementsResponse {
  type?: ResponseType;
  message: string;
  session_id?: SetupRequirementsResponseSessionId;
  setup_info: SetupInfo;
  graph_id?: SetupRequirementsResponseGraphId;
  graph_version?: SetupRequirementsResponseGraphVersion;
}